requirements.txt
tensorflow==2.3.1
pydantic==1.6.1
requests==2.24.0
numpy==1.18.5
uvicorn==0.11.3
Keras==2.4.3
pandas==1.1.1
joblib==0.16.0
fastapi==0.61.2
Pillow==8.0.1
scikit_learn==0.23.2
python-multipart == 0.0.5
aiofiles
jinja2

